Move component 'GroupContentMenu' to folder 'ContentMenu'
