+++ webapp/components/ContentMenu/GroupContentMenu.spec.js
