+++ webapp/components/ContentMenu/GroupContentMenu.vue
